backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.25.0
websockets==13.1
requests==2.32.3
aiohttp==3.10.5
beautifulsoup4==4.12.3
readability-lxml==0.8.1
selenium==4.25.0
playwright==1.48.0
python-multipart==0.0.16
pydantic==2.9.0
structlog==24.4.0
tiktoken==0.8.0